+++ changelog.md
@@ -186,3 +186,4 @@
 * Command kill no longer counts as murder
 * Added config flag to disable name tags
 * Fixed log spam regarding invalid bounding boxes
+* Fixed issues when assigning family in editor